+++ Code/Functions/Public/Write-PSHTMLAsset.ps1
@@ -1,9 +1,23 @@
 function Write-PSHTMLAsset {
     <#
     .SYNOPSIS
-      Add script references to your PSHTML scripts.
+        Add asset references to your PSHTML scripts.
     .DESCRIPTION
         Write-PSHTML will scan the 'Assets' folders in the PSHTML module folder, and 
+        One can Use Write-PSHTML (Without any parameter) to add dynamically a link / script tag for every asset that is available in the Asset Folder.
+
+    .PARAMETER Name
+
+    Specify the name of an Asset. This is a dynamic parameter, and calculates the names based on the content of Assets folder.
+
+    .PARAMETER Type
+
+    Allows to specifiy what type of Asset to return. Script (.js) or Style (.css) are the currently supported ones.
+
+    .EXAMPLE
+        Write-PSHTMLAsset
+
+        Will generate all the asset tags, regardless of their type.
 
     .EXAMPLE
         Write-PSHTMLAsset -Type Script -Name ChartJs
@@ -18,6 +32,12 @@ function Write-PSHTMLAsset {
         Generates the following results:
         
         <Link src='BootStrap/bootstrap.min.css'></Link>
+
+    .EXAMPLE
+        Write-PSHTMLAsset -Name Bootstrap
+
+        Generates all the links regardless of their type.
+
 
     .LINK
         https://github.com/Stephanevg/PSHTML
@@ -42,7 +62,7 @@ function Write-PSHTMLAsset {
  
         $AssetAttribute = New-Object System.Management.Automation.ParameterAttribute
         $AssetAttribute.Position = 2
-        $AssetAttribute.Mandatory = $true
+        $AssetAttribute.Mandatory = $False
         $AssetAttribute.HelpMessage = 'Select the asset to add'
  
         $AssetCollection = New-Object System.Collections.ObjectModel.Collection[System.Attribute]
@@ -64,10 +84,15 @@ function Write-PSHTMLAsset {
     }
     
     process {
-        if($Type){
+        if($Type -and $Name){
             $Asset = (Get-PSHTMLConfiguration).GetAsset($Name,[AssetType]$Type)
-        }Else{
+        }ElseIf($Name){
             $Asset = (Get-PSHTMLConfiguration).GetAsset($Name)
+        }ElseIf($Type){
+            (Get-PSHTMLConfiguration).GetAsset([AssetType]$Type)
+        }
+        Else{
+            $Asset = (Get-PSHTMLConfiguration).GetAsset()
         }
 
         Foreach($A in $Asset){